Start of the executesolver.py modul
The executesolver.py modul was added. At the moment it can only be excuted within the spyder programm. In the future it is supposed to
be excuted from the console. The first output files had been created, but there are some issues with the solver. Some minor changes
for the plotter have been done.